+++ src/commands/mod.rs
@@ -2,3 +2,5 @@ pub mod check;
 pub mod fmt;
 pub mod lint;
 pub mod new;
+pub mod read_properties;
+pub mod to_prompt;